% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.objects.R
\name{second.order.MESH2.5D}
\alias{second.order.MESH2.5D}
\title{Double the order of a fist order Finite Element mesh by adding middle points to each side of the triangles in the triangulation}
\usage{
second.order.MESH2.5D(mesh, bc = NULL)
}
\arguments{
\item{mesh}{an object of class 'SURFACE_MESH' is the starting mesh of order 1}

\item{bc}{A vector specifying the indices of the nodes on which boundary conditions are applied}
}
\value{
if no boundaries conditions are passed
\item{\code{mesh}} An object of class 'SURFACE_MESH' with the mesh of order 2. Otherwhise a \code{\list} with parameters:
\item{\code{mesh}} An object of class 'SURFACE_MESH' with the mesh of order 2.
\item{\code{bc_index}} An update of the vector specifying the indices of the nodes on which boundary conditions are applied
}
\description{
Double the order of a fist order Finite Element mesh by adding middle points to each side of the triangles in the triangulation
}
\examples{
data(hub) #loading mesh hub, order=1
hub_order2 = second.order.MESH2.5D(hub)
}
